Extract viewer rendering and instrumentation into @artifactshare/viewer-kit
Move the markdown document renderer, TanStack highlight setup, and the
framework-free sandbox instrumentation script out of apps/web into a new
workspace package so the CLI preview server can reuse them. Old module
paths in apps/web remain as one-line re-export shims, so no importer
changes. bundle-sandbox re-exports injectReadyReporter unchanged.
